feat(worldgen): add buildPlaza to carve central plaza and wire into town_gen.js
